Remove obsolete code
